requirements.txt
streamlit
joblib
numpy
pandas
scikit-learn
plotly
fpdf2
kaleido

